fix(NotificationSettings.tsx, NotificationsSubscriber.ts): change name
